Minor, move code
